refactor: move `Annotation` from `Ast` to `shared`
